Updated import and organized files after viewer and convert re-organization in scipion repo
