+++ xmipp3/protocols/pdb/protocol_pseudoatoms_base.py
@@ -31,15 +31,12 @@ from os.path import basename, join, abspath
 import os
 
 from pyworkflow.object import String
-from pyworkflow.utils.path import cleanPattern, createLink, moveFile
+from pyworkflow.utils import cleanPattern, createLink, moveFile
 from pyworkflow.protocol.params import EnumParam, PointerParam, FloatParam
 from pyworkflow.protocol.constants import LEVEL_ADVANCED
 from pyworkflow.em.protocol import Prot3D
-from pyworkflow.em.viewers.chimera_utils import \
-    createCoordinateAxisFile
 
 from xmipp3.convert import getImageLocation
-# from xmipp3.base import XmippProtocol
 
 
 NMA_MASK_NONE = 0
