+++ xmipp3/protocols/protocol_compare_angles.py
@@ -24,19 +24,12 @@
 # *
 # **************************************************************************
 
-from os.path import join, exists
-import math
-
 import pyworkflow.protocol.params as params
 from pyworkflow import VERSION_1_2
-from pyworkflow.utils.path import makePath
-from pyworkflow.em.convert import ImageHandler, ALIGN_PROJ
-from pyworkflow.em.data import Image
 from pyworkflow.em.protocol import ProtAnalysis3D
 import pyworkflow.em.metadata as md
 
 import xmippLib
-from xmipp3.base import findRow
 from xmipp3.convert import (rowToAlignment, setXmippAttributes,
                        xmippToLocation, writeSetOfParticles)
 from xmipp3.constants import SYM_URL
@@ -155,7 +148,6 @@ class XmippProtCompareAngles(ProtAnalysis3D):
         
     def _createItemMatrix(self, particle, row):
         setXmippAttributes(particle, row, xmippLib.MDL_SHIFT_DIFF, xmippLib.MDL_ANGLE_DIFF)
-        #createItemMatrix(particle, row, align=em.ALIGN_PROJ)
 
     # --------------------------- INFO functions -------------------------------
