+++ xmipp3/protocols/protocol_extract_unit_cell.py
@@ -26,12 +26,11 @@
 
 from pyworkflow import VERSION_1_2
 from pyworkflow.em import Volume
-from pyworkflow.em.constants import SCIPION_SYM_NAME
 from pyworkflow.em.constants import (SYM_I222, SYM_I222r, SYM_In25, SYM_In25r,
                                      SYM_CYCLIC, SYM_DIHEDRAL, SYM_TETRAHEDRAL,
-                                     SYM_OCTAHEDRAL)
+                                     SYM_OCTAHEDRAL, SCIPION_SYM_NAME)
 from pyworkflow.em.data import Transform
-from pyworkflow.em.headers import Ccp4Header
+from pyworkflow.em.convert import Ccp4Header
 from pyworkflow.em.protocol import EMProtocol
 from pyworkflow.protocol.params import (PointerParam, FloatParam,
                                         EnumParam, IntParam)
