+++ xmipp3/viewers/plotter.py
@@ -27,8 +27,7 @@
 This module implement the classes to create plots on xmipp.
 """
 
-#from pyworkflow.gui.plotter import Plotter
-from pyworkflow.em.plotter import EmPlotter
+from pyworkflow.em.viewers.plotter import EmPlotter
 
 
 class XmippPlotter(EmPlotter):
