+++ xmipp3/viewers/viewer_movie_alignment.py
@@ -25,7 +25,7 @@
 # **************************************************************************
 
 from pyworkflow.viewer import Viewer, DESKTOP_TKINTER, WEB_DJANGO
-import pyworkflow.em.showj as showj
+import pyworkflow.em.viewers.showj as showj
 
 from xmipp3.protocols.protocol_movie_opticalflow import (XmippProtOFAlignment,
                                                  OBJCMD_MOVIE_ALIGNCARTESIAN)
