+++ xmipp3/viewers/viewer_nma.py
@@ -31,11 +31,11 @@ visualization program.
 from pyworkflow.gui.project import ProjectWindow
 from pyworkflow.protocol.params import LabelParam, IntParam
 from pyworkflow.viewer import ProtocolViewer, DESKTOP_TKINTER, WEB_DJANGO
-from pyworkflow.em.viewer import ObjectView, VmdView
+from pyworkflow.em.viewers import ObjectView, VmdView
 
 import xmippLib
-from .protocol_nma import XmippProtNMA
-from .plotter import XmippNmaPlotter
+from xmipp3.protocols import XmippProtNMA
+from .nma_plotter import XmippNmaPlotter
 
 
 OBJCMD_NMA_PLOTDIST = "Plot distance profile"
