+++ xmipp3/viewers/viewer_nma_dimred.py
@@ -37,13 +37,13 @@ from pyworkflow.viewer import (ProtocolViewer, DESKTOP_TKINTER, WEB_DJANGO)
 from pyworkflow.protocol.params import StringParam, LabelParam
 from pyworkflow.em.data import SetOfParticles
 from pyworkflow.utils.process import runJob
-from pyworkflow.em.viewer import VmdView
+from pyworkflow.em.viewers import VmdView
 from pyworkflow.gui.browser import FileBrowserWindow
 
-from .protocol_nma_dimred import XmippProtDimredNMA, DIMRED_MAPPINGS
-from .data import Point, Data
-from .plotter import XmippNmaPlotter
-from .gui import ClusteringWindow, TrajectoriesWindow
+from xmipp3.protocols.nma.protocol_nma_dimred import XmippProtDimredNMA, DIMRED_MAPPINGS
+from xmipp3.protocols.nma.data import Point, Data
+from .nma_plotter import XmippNmaPlotter
+from .nma_gui import ClusteringWindow, TrajectoriesWindow
 
         
 class XmippDimredNMAViewer(ProtocolViewer):
