+++ xmipp3/viewers/viewer_normalize_strain.py
@@ -25,7 +25,7 @@
 # **************************************************************************
 
 from pyworkflow.viewer import ProtocolViewer, DESKTOP_TKINTER, WEB_DJANGO
-from pyworkflow.em.viewer import DataView, ChimeraView
+from pyworkflow.em.viewers import DataView, ChimeraView
 from xmipp3.viewers import XmippViewer
 from pyworkflow.protocol.params import PointerParam, LabelParam
