+++ xmipp3/viewers/viewer_projmatch.py
@@ -34,18 +34,15 @@ import numpy as np
 
 from pyworkflow.protocol.executor import StepExecutor
 from pyworkflow.viewer import CommandView, Viewer, ProtocolViewer, DESKTOP_TKINTER, WEB_DJANGO
-from pyworkflow.em.viewer import DataView, ClassesView, Classes3DView
+from pyworkflow.em.viewers import DataView, ClassesView, Classes3DView, EmPlotter
 from pyworkflow.utils import createUniqueFileName, cleanPattern
-from pyworkflow.em.viewer import ChimeraDataView
 from pyworkflow.protocol.constants import LEVEL_ADVANCED
 from pyworkflow.protocol.params import (LabelParam, IntParam, FloatParam,
                                         StringParam, EnumParam, NumericRangeParam)
-from pyworkflow.em.plotter import EmPlotter
-
 import xmippLib
 from xmipp3.convert import *
 from xmipp3.viewers.plotter import XmippPlotter
-from .protocol_projmatch import XmippProtProjMatch
+from xmipp3.protocols import XmippProtProjMatch
 # from .projmatch_initialize import createFilenameTemplates
