+++ xmipp3/viewers/viewer_volume_strain.py
@@ -28,7 +28,7 @@
 # **************************************************************************
 
 from pyworkflow.viewer import ProtocolViewer, DESKTOP_TKINTER, WEB_DJANGO
-from pyworkflow.em.viewer import DataView, ChimeraView
+from pyworkflow.em.viewers import DataView, ChimeraView
 
 from .viewer import XmippViewer
 from xmipp3.protocols.protocol_volume_strain import XmippProtVolumeStrain
